Nav Bar finalized, footer cleaned up
I finished the navigation bar/ header. I adjusted the canvas size on the
contact icon to avoid the annoying padding/margin issue. Added a
complimentary background image to the footer and cleaned up some of the
text colors and positions. Copied the header/footer to the other pages
and adjusted the "currentPage" class designation Edited several of the
icons and images in GIMP.